added slack routes
